apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "SportsLab FastAPI backend (multi-tenant, Clerk-authenticated)."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "SportsLab" },
]

# Runtime dependencies. Keep minimal — this is the edge service and every
# extra dep slows cold-start and increases the attack surface.
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic[email]>=2.8,<3.0",  # [email] extra pulls email-validator (EmailStr, SPO-127)
    "pydantic-settings>=2.4,<3.0",
    "httpx>=0.27,<1.0",             # Clerk JWKS fetch
    "python-jose[cryptography]>=3.3,<4.0",  # JWT verification against Clerk JWKS
    "structlog>=24.4,<25.0",
    "sqlalchemy>=2.0,<3.0",
    "svix>=1.24,<2.0",              # Clerk webhook signature verification (Svix standard)
    "stripe>=10.0,<12.0",           # Stripe SDK + webhook signature verification
    # Needed for ``ZoneInfo("Europe/Warsaw")`` on Windows hosts and on
    # slim Docker images where the OS tzdata is absent. Tiny (~400 kB)
    # and a runtime hard dep because ``predictions.py`` requires it at import.
    "tzdata>=2024.1",
]

[project.optional-dependencies]
# Database drivers are optional — API can run in dry-run / SQLite mode for
# local development and tests. Production deployments install [postgres].
postgres = [
    "asyncpg>=0.29,<1.0",
]
sqlite = [
    "aiosqlite>=0.20,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27,<1.0",  # TestClient uses httpx under the hood
    "types-python-jose>=3.3,<4.0",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build]
# See packages/ml-in-sports/pyproject.toml — root .gitignore excludes `models/`
# which would incorrectly drop src/api/models/ from the wheel.
ignore-vcs = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.hatch.build.targets.sdist]
include = [
    "src/api",
    "tests",
    "README.md",
    "pyproject.toml",
]
